Update MAX32625PICO Bootloader
BL image generated from below SHA point
https://github.com/ARMmbed/DAPLink/commit/82dc31d09456710ad0610db5b0ee2a73282f6acc

Detail:
When code attempted to release the ports by placing them into Hi-Z mode,
it would first clear the out register (so that weak pull-up would be disabled)
then made the switch to Hi-Z. Since the signals were initially in output mode,
clearing the out register caused a low pulse on the pins. One of the pins is
the RESETN line which caused an unwanted device reset.

MAX32625PICO BL gpio is going to to use the "NORMAL_HI_Z" mode
that the MAX32625 supports.
This mode places the pins in Hi-Z (no pull-up) regardless of
the state of the out register.

PR: https://github.com/ARMmbed/DAPLink/pull/1074

Signed-off-by: Sadik Ozer <sadik.ozer@analog.com>